(Front) [Fix] Move View Component into views Folder_1
작업 내용
1. view에 해당하는 component를 views folder에 이동 : test
  - 정상적으로 실행됨을 확인
