+++ frontend/src/router/index.js
@@ -1,5 +1,5 @@
 import { createRouter, createWebHistory } from "vue-router";
-import HomePage from "../components/client/common/home/pages/HomePage.vue";
+import HomePage from "@/views/HomePage.vue";
 
 const routes = [
   /* 공통 */
